Extra hooks code to own module
